Conflicts:
	Code/Library/communication/state.c
	Code/Library/communication/state.h
	Code/Library/control/stabilisation_copter.c
	Code/Library/control/stabilisation_copter.h
	Code/Library/hal/remote_controller.h
	Code/Library/sensing/simulation.c
	Code/Library/sensing/simulation.h
	Code/Maveric_myCopter/Maveric_myCopter.cproj
	Code/Maveric_myCopter/Maveric_myCopter_linux.project
	Code/Maveric_myCopter/src/central_data.c
	Code/Maveric_myCopter/src/central_data.h
	Code/Maveric_myCopter/src/tasks.c

+++ Code/Library/util/buffer.c
@@ -28,7 +28,7 @@ uint8_t buffer_full(buffer_t * buffer)
 
 uint8_t buffer_put_lossy(buffer_t * buffer, uint8_t byte) 
 {
-	uint16_t tmp;
+	uint8_t tmp;
 	tmp = (buffer->buffer_head + 1)&BUFFER_MASK;
 
 	if (tmp==buffer->buffer_tail) 
@@ -57,7 +57,7 @@ uint8_t buffer_put_lossy(buffer_t * buffer, uint8_t byte)
 
 uint8_t buffer_put(buffer_t * buffer, uint8_t byte) 
 {
-	uint16_t tmp;
+	uint8_t tmp;
 	tmp = (buffer->buffer_head + 1)&BUFFER_MASK;
 
 	if (tmp==buffer->buffer_tail)
